requirements.txt
Django==1.8.7
django-rest-swagger==0.3.4
djangorestframework==3.3.1
MySQL-python==1.2.5
PyYAML==3.11
wheel==0.24.0
